requirements.txt
environs==14.1.1
vk-api==11.9.9
google-cloud-dialogflow==2.41.1
google-cloud-api-keys==0.5.16
python-telegram-bot==13.15
urllib3==1.26.15
standard-imghdr==3.13.0
